backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
PyPDF2==3.0.1
google-generativeai==0.3.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiosmtplib==3.0.1
aiofiles==23.2.1
